+++ modules/addons/ingress-nginx/variables.tf
@@ -4,6 +4,12 @@ variable "loadbalancer_ip" {
   default     = null
 }
 
+variable "namespace" {
+  description = "Kubernetes namespace for ingress-nginx"
+  type        = string
+  default     = "ingress-nginx"
+}
+
 variable "chart_version" {
   description = "Ingress-NGINX Helm chart version"
   type        = string
@@ -26,4 +32,37 @@ variable "default_ingress_class" {
   description = "Make this the default ingress class"
   type        = bool
   default     = true
+}
+
+variable "controller_replicas" {
+  description = "Number of controller replicas"
+  type        = number
+  default     = 1
+}
+
+variable "controller_resources" {
+  description = "Resource requests/limits for ingress controller"
+  type = object({
+    requests = optional(object({
+      cpu    = optional(string, "100m")
+      memory = optional(string, "128Mi")
+    }), {})
+    limits = optional(object({
+      cpu    = optional(string, "500m")
+      memory = optional(string, "512Mi")
+    }), {})
+  })
+  default = {}
+}
+
+variable "enable_metrics" {
+  description = "Enable Prometheus metrics"
+  type        = bool
+  default     = false
+}
+
+variable "metrics_port" {
+  description = "Port for Prometheus metrics"
+  type        = number
+  default     = 10254
 }